+++ lm_engine/__init__.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2026, Mayank Mishra
 # **************************************************
 
-from .register_hf import register_model_classes
+from .hf_adapter import register_model_classes
 
 
 register_model_classes()
