+++ lm_engine/model_wrapper/base.py
@@ -14,13 +14,12 @@ from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
 
 from ..dtensors import tensor_to_dtensor
 from ..enums import Kernel
-from ..hf_adapter import LLMAdapter_HF
+from ..hf_adapter import LLMAdapter_HF, get_causal_lm_class, is_custom_model
 from ..kernels import is_kernel_allowed
 from ..logging_utils import log_rank_0
 from ..loss import get_autoregressive_language_modeling_loss, is_aux_loss_zero
 from ..modeling_utils import CausalLMOutputWithPast
 from ..parallel import ProcessGroupManager
-from ..register_hf import get_causal_lm_class, is_custom_model
 from ..tokenizers import get_tokenizer
 from ..utils import SafeTensorsWeightsManager, string_to_torch_dtype
