+++ lm_engine/train_utils.py
@@ -8,11 +8,11 @@ import torch
 from torch.distributed import ReduceOp
 
 from .enums import GradientCheckpointingMethod
+from .hf_adapter import is_custom_model
 from .logging_utils import ExperimentsTracker, MetricsTrackingDict, log_metrics
 from .model_config import CommonConfig
 from .modeling_utils import is_glu
 from .parallel import ProcessGroupManager
-from .register_hf import is_custom_model
 
 
 def all_reduce_metrics_tracker(metrics_tracker: MetricsTrackingDict) -> MetricsTrackingDict:
